backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pymongo>=4.6.0
dnspython>=2.6.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
PyJWT>=2.8.0
python-dateutil>=2.9.0
openpyxl>=3.1.0
pillow>=10.0.0
numpy>=1.26.0,<2.1
opencv-python-headless>=4.9.0
deepface>=0.0.92
